Crate fedimint_connectors

Modules

error

http

iroh

tor

ws

Structs

ConnectionPool

ConnectionState

ConnectionStateInner 🔒

Inner part of ConnectionState preserving state between attempts to initialize ConnectionState::connection

ConnectorRegistry

A set of available connectivity protocols a client can use to make network API requests (typically to federation).

ConnectorRegistryBuilder

Builder for ConnectorRegistry

ConnectorRegistryInner 🔒

Actual data shared between copies of ConnectorRegistry handle

Traits

Connector

IConnection

Generic connection trait shared between IGuardianConnection and IGatewayConnection

IGatewayConnection

A connection from a client to a gateway

IGuardianConnection

A connection from api client to a federation guardian

Type Aliases

ConnectorInitFn 🔒

Type for connector initialization functions

DynConnector

DynGatewayConnection

A connection from api client to a gateway (type erased)

DynGuaridianConnection

A connection from api client to a federation guardian (type erased)

ServerResult